+++ changelog.md
@@ -10,6 +10,7 @@ Codion Change Log
 ### is.codion.swing.fraework.ui
 - EntitySearchField.searchControl() added.
 - EntitySearchFieldPanel, search button added.
+- EntityPanel, EntityEditPanel, EntityTablePanel.Config now provides access the the panel being configured.
 
 ## 0.17.39
 ### is.codion.common.core